fix: organized files into globals folder
