+++ client/src/AdminPanelGuard.jsx
@@ -3,7 +3,7 @@ import { Outlet, Navigate } from "react-router-dom";
 import { useUserDetails } from "./context/UserContext";
 import { USER_ROLE } from "./core/utils/const";
 import { useUserAuth } from "./context/UserAuthContext";
-import NewNavbar from "./components/NewNavbar";
+import NewNavbar from "./components/globals/NewNavbar";
 
 function AdminPanelGuard() {
   const { userDetails } = useUserDetails();
