+++ client/src/components/CustomComponents.jsx
@@ -1,26 +1,26 @@
-import BgEllipse from "./BgEllipse";
-import Button from "./Button";
-import Card from "./Card";
+import BgEllipse from "./globals/BgEllipse";
+import Button from "./globals/Button";
+import Card from "./globals/Card";
 import Contests from "./Contests";
 import About from "./About";
 import CopyToClipboard from "./CopyToClipboard";
-import ErrorPage from "./error-page";
+import ErrorPage from "./globals/error-page";
 import Feedback from "./Feedback";
 import Home from "./Home";
-import IndividualCard from "./IndividualCard";
+import IndividualCard from "./globals/IndividualCard";
 import Layout from "./Layout";
 import Login from "./Login";
-import Logout from "./Logout";
-import MobNav from "./MobNav";
-import Navbar from "./Navbar";
-import Homepage from "./Homepage";
-import Footer from "./Footer";
+import Logout from "./globals/Logout";
+import MobNav from "./globals/MobNav";
+import Navbar from "./globals/Navbar";
+import Homepage from "../pages/home/Homepage";
+import Footer from "./globals/Footer";
 import NewLogin from "./NewLogin";
 import { loader as loginLoader } from "./Login";
-import { loader as signupLoader } from "./Signup";
-import NewNavbar from "./NewNavbar";
-import ScrollToTop from "./ScrollToTop";
-import Signup from "./Signup";
+import { loader as signupLoader } from "./globals/Signup";
+import NewNavbar from "./globals/NewNavbar";
+import ScrollToTop from "./globals/ScrollToTop";
+import Signup from "./globals/Signup";
 import Updates from "./Updates";
 export {
   BgEllipse,
