+++ client/src/components/globals/Logout.jsx
@@ -2,7 +2,7 @@ import { useEffect } from "react";
 
 import { useNavigate } from "react-router-dom";
 
-import { auth } from "../../firebase";
+import { auth } from "../../../firebase";
 
 import { toast } from "react-toastify";
