+++ client/src/components/globals/Signup.jsx
@@ -7,17 +7,17 @@ import {
 } from "react-router-dom";
 import axios from "axios";
 import { useState } from "react";
-import { isLoggedIn } from "../../api";
-import { auth } from "../../firebase";
-import { useUserAuth } from "../context/UserAuthContext";
+import { isLoggedIn } from "../../../api";
+import { auth } from "../../../firebase";
+import { useUserAuth } from "../../context/UserAuthContext";
 
-import GoogleAuthButton from "./AuthButtons/GoogleAuthButton";
-import GithubAuthButton from "./AuthButtons/GithubAuthButton";
+import GoogleAuthButton from "../AuthButtons/GoogleAuthButton";
+import GithubAuthButton from "../AuthButtons/GithubAuthButton";
 import { ToastContainer, toast } from "react-toastify";
-import loginIcon from "../assets/fingerprint-animate-blue.svg";
+import loginIcon from "/src/assets/fingerprint-animate-blue.svg";
 
 const backendUrl = import.meta.env.VITE_REACT_APP_BACKEND_URL;
-import SignoutButton from "../user/components/SignoutButton";
+import SignoutButton from "../../user/components/SignoutButton";
 
 export async function loader() {
   const loggedIn = await isLoggedIn();
