+++ client/src/main.jsx
@@ -6,7 +6,7 @@ import App from "./App";
 
 import "./index.css";
 import "tw-elements-react/dist/css/tw-elements-react.min.css";
-import InstallPWAButton from "./components/InstallPWAButton";
+import InstallPWAButton from "/src/components/globals/InstallPWAButton";
 
 import { Helmet } from "react-helmet";
 import metaData from "./metaData.json";
